requirements.txt
prisma == 0.12.0
gunicorn
fastapi
uvicorn
python-dotenv
python-jose
passlib
python-multipart
bcrypt == 4.0.1